feat: hidden Textarea 크기를 정하는 코드 hook으로 관리하고, 쓸데없이 MemoCreateTextArea에서 구현한 이미지 파일 불러오는 기능을 IconButtons 컴포넌트로 이동
