+++ src/pages/home/subPages/main/components/MemoCreateTextArea/MemoCreateTextArea.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, useContext, useEffect, useRef, useState } from 'react';
+import { ChangeEvent, useContext, useRef, useState } from 'react';
 import { TextareaAutosize } from '@mui/material';
 import { usePressEnterFetch } from 'pages/home/subPages/hooks';
 import { IconButtons } from './IconButtons';
@@ -6,6 +6,7 @@ import { HiddenTextarea } from './HiddenTextarea';
 import { ImageList } from './ImageList';
 import { ImageListContext } from 'utils';
 import { isFilesResponse, uploadFile, uploadFiles } from 'api';
+import { useHiddenTextareaManager } from './hook';
 
 interface MemoCreateTextAreaProps {
   value: string;
@@ -23,14 +24,10 @@ const MemoCreateTextArea = ({
   const { images, addImage, removeImage, removeAllImage } =
     useContext(ImageListContext);
   const [focus, setFocus] = useState(false);
-  const [isMultiline, setIsMultiline] = useState(false);
-  const [hiddenTextareaWidth, setHiddenTextareaWidth] = useState<number | null>(
-    null
-  );
-
-  const hiddenTextareaRef = useRef<HTMLTextAreaElement>(null);
   const containerRef = useRef<HTMLDivElement>(null);
 
+  const { hiddenTextareaRef, isMultiline, hiddenTextareaWidth } =
+    useHiddenTextareaManager(value, images);
   const { handlePressEnterFetch } = usePressEnterFetch({
     handleEnterWithCtrl: handleSubmit,
   });
@@ -78,33 +75,6 @@ const MemoCreateTextArea = ({
     // TODO: 마이크 버튼 클릭시 하는 메소드 생기면 추가
   };
 
-  const handleFileChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const files = e.target.files;
-    if (files && files.length > 0) {
-      addImage(files[0]);
-    }
-  };
-
-  const handleCameraButtonClick = () => {
-    const inputFile = document.querySelector(
-      'input[type="file"]'
-    ) as HTMLInputElement;
-    if (inputFile) inputFile.click();
-  };
-
-  useEffect(() => {
-    if (hiddenTextareaRef.current && containerRef.current) {
-      const lineHeight = parseFloat(
-        getComputedStyle(hiddenTextareaRef.current).lineHeight
-      );
-      const maxLines = Math.floor(
-        hiddenTextareaRef.current.clientHeight / lineHeight
-      );
-      setIsMultiline(maxLines > 1 || images.length >= 1);
-      setHiddenTextareaWidth(containerRef.current.clientWidth);
-    }
-  }, [value, images]);
-
   return (
     <div
       className="flex flex-shrink-0 px-4 py-3 rounded-2xl overflow-hidden gap-4 bg-[#FFF6E3CC] border border-black border-opacity-10 font-regular shadow-custom backdrop-blur-lg"
@@ -136,8 +106,6 @@ const MemoCreateTextArea = ({
         <IconButtons
           submitAvailable={focus || isMultiline}
           onMicButtonClick={handleMicButtonClick}
-          onCameraButtonClick={handleCameraButtonClick}
-          onFileChange={handleFileChange}
           onSubmitButtonClick={handleSubmit}
         />
       </div>
