+++ src/pages/home/subPages/main/components/MemoCreateTextArea/hook/index.ts
@@ -1,0 +1,1 @@
+export { default as useHiddenTextareaManager } from './useHiddenTextareaManager';
